requirements.txt
absl-py==2.3.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
attrs==25.3.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
easyocr==1.7.2
faiss-cpu==1.11.0
filelock==3.18.0
filterpy==1.4.5
flake8==7.3.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mediapipe==0.10.21
ml_dtypes==0.5.1
motmetrics==1.4.0
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==1.44.0
networkx==3.5
ninja==1.11.1.4
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==80.9.0
shapely==2.1.1
simsimd==6.4.9
six==1.17.0
sounddevice==0.5.2
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.16
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.160
ultralytics-thop==2.0.14
urllib3==2.5.0
xmltodict==0.14.2
